refactor(conversationstore,failover): serve both SQL backends from one store
These two are where the dialects genuinely diverge, so the split is kept where
it is real and removed everywhere else.

conversationstore's MergeMetadata, AppendItems and DeleteItem must mutate JSON
server-side in a single UPDATE or two concurrent Responses turns overwrite each
other's exchange. SQLite expresses that with JSON1 functions, PostgreSQL with
jsonb operators, and there is no portable spelling. The statements now live
side by side behind a jsonMutations interface while the surrounding logic —
rows-affected interpretation, duplicate detection, re-read — is shared. Their
tests run on both engines for the first time.

failover's legacy rename is likewise irreducible: PostgreSQL renames columns in
place, SQLite copies rows into a reshaped table (which is also where padded
primary keys get trimmed). Both are kept verbatim behind a dialect switch
rather than rewritten, since they run against databases in the field.

The migration test now covers both engines. The PostgreSQL migration had no
test at all before this — the survey found the two halves had drifted into
independently written trim logic with only one of them verified.
